Move IRHandle and TaskLaunchRecord into async_utils.h
