+++ taichi/program/async_utils.h
@@ -1,10 +1,71 @@
 #pragma once
 
+#include <unordered_map>
 #include <unordered_set>
 
 #include "taichi/ir/snode.h"
+#define TI_RUNTIME_HOST
+#include "taichi/program/context.h"
+#undef TI_RUNTIME_HOST
 
 TLANG_NAMESPACE_BEGIN
+
+class IRNode;
+
+class IRHandle {
+ public:
+  IRHandle(const IRNode *ir, uint64 hash) : ir_(ir), hash_(hash) {
+  }
+
+  std::unique_ptr<IRNode> clone() const;
+
+  const IRNode *ir() const {
+    return ir_;
+  }
+
+  uint64 hash() const {
+    return hash_;
+  }
+
+  // Two IRHandles are considered the same iff their hash values are the same.
+  bool operator==(const IRHandle &other_ir_handle) const {
+    return hash_ == other_ir_handle.hash_;
+  }
+
+ private:
+  const IRNode *ir_;  // not owned
+  uint64 hash_;
+};
+
+TLANG_NAMESPACE_END
+
+namespace std {
+template <>
+struct hash<taichi::lang::IRHandle> {
+  std::size_t operator()(const taichi::lang::IRHandle &ir_handle) const
+      noexcept {
+    return ir_handle.hash();
+  }
+};
+}  // namespace std
+
+TLANG_NAMESPACE_BEGIN
+
+class OffloadedStmt;
+
+// Records the necessary data for launching an offloaded task.
+class TaskLaunchRecord {
+ public:
+  Context context;
+  Kernel *kernel;  // TODO: remove this
+  IRHandle ir_handle;
+
+  TaskLaunchRecord();
+
+  TaskLaunchRecord(Context context, Kernel *kernel, IRHandle ir_handle);
+
+  inline OffloadedStmt *stmt() const;
+};
 
 struct AsyncState {
   enum class Type { mask, value, list };
