+++ taichi/program/state_flow_graph.cpp
@@ -16,8 +16,9 @@ StateFlowGraph::StateFlowGraph() {
   initial_node_->launch_id = 0;
 }
 
-void StateFlowGraph::insert_task(const TaskMeta &task_meta) {
+void StateFlowGraph::insert_task(const TaskLaunchRecord &rec, const TaskMeta &task_meta) {
   auto node = std::make_unique<Node>();
+  node->rec = rec;
   node->task_name = task_meta.kernel_name;
   {
     int &id = task_name_to_launch_ids_[node->task_name];
